Expand & organize integration tests
