+++ tests/integration/test_cli_comparison.py
